requirements.txt
aioredis==1.3.1
argcomplete==1.9.3
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
boto==2.49.0
boto3==1.11.10
botocore==1.14.10
certifi==2019.11.28
cffi==1.13.2
cfn-flip==1.2.2
channels==2.3.0
channels-redis==2.3.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.1
django-cors-headers==3.2.1
django-dynamodb-sessions==0.6
docutils==0.16
durationpy==0.5
future==0.16.0
hiredis==1.0.1
hjson==3.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
msgpack==0.5.6
placebo==0.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
s3transfer==0.3.2
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.3
Twisted==19.10.0
txaio==20.1.1
Unidecode==1.1.1
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.48.2
zope.interface==4.7.1